CardanoClient

Description

A basic cardano-node client that can talk to a local cardano-node.

The idea of this module is to provide a Haskell interface on top of cardano-cli's API, using cardano-api types.

Documentation

type NodeSocket = FilePath

buildAddress :: VerificationKey PaymentKey -> NetworkId -> Address ShelleyAddr

Build an address give a key.

From https://github.com/input-output-hk/cardano-node/blob/master/cardano-cli/src/Cardano/CLI/Shelley/Run/Address.hs#L106 Throws CardanoClientException if the query fails.

queryUtxo :: NetworkId -> FilePath -> [Address ShelleyAddr] -> IO (UTxO AlonzoEra)

Query UTxO for all given addresses.

This query is specialised for Shelley addresses in Alonzo era. Throws CardanoClientException if query fails.

txOutLovelace :: TxOut era -> Lovelace

Extract ADA value from an output NOTE(AB): there is txOutValueToLovelace in more recent cardano-api versions which serves same purpose

queryProtocolParameters :: NetworkId -> FilePath -> IO ProtocolParameters

Query current protocol parameters.

Throws CardanoClientException if query fails.

runQuery :: NetworkId -> FilePath -> QueryInMode CardanoMode (Either EraMismatch a) -> IO a

localNodeConnectInfo :: NetworkId -> FilePath -> LocalNodeConnectInfo CardanoMode

cardanoModeParams :: ConsensusModeParams CardanoMode

queryTipSlotNo :: NetworkId -> FilePath -> IO SlotNo

buildRaw :: [TxIn] -> [TxOut AlonzoEra] -> SlotNo -> Lovelace -> Either TxBodyError (TxBody AlonzoEra)

Build a "raw" transaction from a bunch of inputs, outputs and fees.

calculateMinFee :: NetworkId -> TxBody AlonzoEra -> Sizes -> ProtocolParameters -> Lovelace

data Sizes

Constructors

Sizes

Fields

inputs :: Int

outputs :: Int

witnesses :: Int

defaultSizes :: Sizes

sign :: SigningKey -> TxBody AlonzoEra -> Tx AlonzoEra

Sign a transaction body with given signing key.

submit :: NetworkId -> FilePath -> Tx AlonzoEra -> IO ()

Submit a (signed) transaction to the node.

Throws CardanoClientException if submission fails.

data CardanoClientException

Constructors

BuildAddressException Text

QueryException Text

BuildRawException TxBodyError

SubmitException Text