CardanoNode

Documentation

type Port = Int

newtype NodeId

Constructors

NodeId Int

data RunningNode

Constructors

RunningNode NodeId FilePath

data CardanoNodeConfig

Configuration parameters for a single node of the cluster

Constructors

CardanoNodeConfig

Fields

nodeId :: NodeId

An identifier for the node

stateDirectory :: FilePath

Parent state directory in which create a state directory for the cluster

systemStart :: UTCTime

Blockchain start time

ports :: PortsConfig

A list of port

data CardanoNodeArgs

Arguments given to the 'cardano-node' command-line to run a node.

Constructors

CardanoNodeArgs

Fields

nodeSocket :: FilePath

nodeConfigFile :: FilePath

nodeByronGenesisFile :: FilePath

nodeShelleyGenesisFile :: FilePath

nodeTopologyFile :: FilePath

nodeDatabaseDir :: FilePath

nodeDlgCertFile :: Maybe FilePath

nodeSignKeyFile :: Maybe FilePath

nodeOpCertFile :: Maybe FilePath

nodeKesKeyFile :: Maybe FilePath

nodeVrfKeyFile :: Maybe FilePath

nodePort :: Maybe Port

defaultCardanoNodeArgs :: CardanoNodeArgs

data PortsConfig

Configuration of ports from the perspective of a peer in the context of a fully sockected topology.

Constructors

PortsConfig

Fields

ours :: Port

Our node TCP port.

peers :: [Port]

Other peers TCP ports.

withCardanoNode :: Tracer IO NodeLog -> CardanoNodeConfig -> CardanoNodeArgs -> (RunningNode -> IO a) -> IO a

cardanoNodeProcess :: Maybe FilePath -> CardanoNodeArgs -> CreateProcess

Generate command-line arguments for launching cardano-node.

refreshSystemStart :: CardanoNodeConfig -> CardanoNodeArgs -> IO ()

Re-generate configuration and genesis files with fresh system start times.

mkTopology :: [Port] -> Value

Generate a topology file from a list of peers.

cliCreateProcess

Arguments

:: Tracer IO NodeLog
for logging the command

-> FilePath
cardano node socket path

-> [Text]
command-line arguments

-> IO CreateProcess

Make a CreateProcess for running cardano-cli. The program must be on the PATH, as normal. Sets CARDANO_NODE_SOCKET_PATH for the subprocess, if a CardanoNodeConn is provided.

data ChainTip

Constructors

ChainTip

Fields

slot :: Integer

hash :: Text

block :: Integer

cliQueryTip

Arguments

:: Tracer IO NodeLog

-> FilePath
cardano node socket path

-> IO ChainTip

Query a cardano node tip with retrying.

cliRetry

Arguments

:: Tracer IO NodeLog

-> Text
message to print before running command

-> CreateProcess

-> IO ByteString

Runs a cardano-cli command and retries for up to 30 seconds if the command failed.

Assumes cardano-cli is available in PATH.

data ProcessHasExited

Constructors

ProcessHasExited Text ExitCode

data NodeLog

Constructors

MsgNodeCmdSpec CmdSpec

MsgCLI [Text]

MsgCLIStatus Text ExitCode

MsgCLIRetry Text

MsgCLIRetryResult Text Int

MsgSocketIsReady FilePath

addField :: ToJSON a => Text -> a -> Value -> Value

withObject :: (Object -> Object) -> Value -> Value

Do something with an a JSON object. Fails if the given JSON value isn't an object.

unsafeDecodeJsonFile :: FromJSON a => FilePath -> IO a