Hydra.Generator

Documentation

data Dataset

A Dataset that can be run for testing purpose. The transactionSequence is guaranteed to be applicable, in sequence, to the initialUtxo set.

Constructors

Dataset

Fields

initialUtxo :: Utxo CardanoTx

transactionsSequence :: [CardanoTx]

generateDataset :: Int -> IO Dataset

Generate an arbitrary UTXO set and a sequence of transactions for this set.

generateConstantUtxoDataset :: Int -> IO Dataset

Generate a Dataset which does not grow the UTXO set over time.

genConstantUtxoDataset :: Int -> Gen Dataset

mkCredentials :: Int -> CardanoKeyPair

generateWith :: Gen a -> Int -> a