HydraNode

Documentation

data HydraClient

Constructors

HydraClient

Fields

hydraNodeId :: Int

connection :: Connection

tracer :: Tracer IO EndToEndLog

withHydraNode :: forall alg. DSIGNAlgorithm alg => Tracer IO EndToEndLog -> FilePath -> (Int, Int, Int) -> Int -> SignKeyDSIGN alg -> [VerKeyDSIGN alg] -> [Int] -> (HydraClient -> IO ()) -> IO ()

send :: HydraClient -> Value -> IO ()

input :: Text -> [Pair] -> Value

Create an input as expected by send.

waitFor :: HasCallStack => Tracer IO EndToEndLog -> Natural -> [HydraClient] -> Value -> IO ()

Wait some time for a single output from each of given nodes. This function waits for delay seconds for message expected to be seen by all given nodes.

waitMatch :: HasCallStack => Natural -> HydraClient -> (Value -> Maybe a) -> IO a

output :: Text -> [Pair] -> Value

Create an output as expected by waitFor and waitForAll.

getMetrics :: HasCallStack => HydraClient -> IO ByteString

queryNode :: Int -> IO (Response ByteString)

defaultArguments :: Int -> FilePath -> [FilePath] -> (Int, Int, Int) -> [Int] -> [String]

withMockChain :: ((Int, Int, Int) -> IO ()) -> IO ()

hydraNodeProcess :: [String] -> CreateProcess

proc :: FilePath -> [String] -> CreateProcess

readCreateProcess :: CreateProcess -> String -> IO String

waitForProcess :: ProcessHandle -> IO ExitCode

withCreateProcess :: CreateProcess -> (Maybe Handle -> Maybe Handle -> Maybe Handle -> ProcessHandle -> IO a) -> IO a

data CreateProcess

Constructors

CreateProcess

Fields

cmdspec :: CmdSpec

cwd :: Maybe FilePath

env :: Maybe [(String, String)]

std_in :: StdStream

std_out :: StdStream

std_err :: StdStream

close_fds :: Bool

create_group :: Bool

delegate_ctlc :: Bool

detach_console :: Bool

create_new_console :: Bool

new_session :: Bool

child_group :: Maybe GroupID

child_user :: Maybe UserID

use_process_jobs :: Bool

data ProcessHandle

data StdStream

Constructors

Inherit

UseHandle Handle

CreatePipe

NoStream

waitForNodesConnected :: HasCallStack => Tracer IO EndToEndLog -> [Int] -> [HydraClient] -> IO ()

withTempDir :: String -> (FilePath -> IO r) -> IO r

Create a temporary directory for the given action to use. The directory is removed if and only if the action completes successfuly.

waitNext :: HasCallStack => HydraClient -> IO Value

withNewClient :: HydraClient -> (HydraClient -> IO a) -> IO a

Runs an action with a new connection to given Hydra node.

withHydraCluster :: Tracer IO EndToEndLog -> FilePath -> (Int, Int, Int) -> Word64 -> (NonEmpty HydraClient -> IO ()) -> IO ()